+++ dotcom-rendering/src/lib/worldCup2026.ts
@@ -1,0 +1,7 @@
+export const worldCup2026PageIds = [
+	'football/world-cup-2026',
+	'football/world-cup-2026/fixtures',
+	'football/world-cup-2026/overview',
+];
+
+export const worldCupTagId = 'football/world-cup-2026';
